+++ extension/src/fg-config-files.ts
@@ -1,9 +1,0 @@
-export const FLAVOR_GRENADE_CONFIG_FILES = [
-  '.fgignore',
-  '.fgattributes',
-] as const;
-
-export const FLAVOR_GRENADE_CONFIG_GLOBS = [
-  '**/.fgignore',
-  '**/.fgattributes',
-] as const;